Break up module statements
The statements module contained a lot of other modules that did not
need the nesting. The modules basic_types, constraints, and
solve_items were promoted to the top level.

For parameters, predicates, and variables their corresponding
_declarations and _types module were moved to be in modules
{parameters,predicates,variables}/{declarations,types}.rs
instead. This gives a similar structure to all three modules for
simlplicity.

The draw-back of the latter change is that to import two types.rs
modules, renaming the module is needed, for example (from
predicates/types.rs)

    use crate::parameters::types as parameter_types;
    use crate::variables::types as variable_types;