Handle whitespace around comma in array definition.
The array `[x1, x2]` and `[x1,x2]` (note the whitespace) should be the
same. However, the former caused the parser to fail. The fix is to parse
whitespace preceding and succeeding the comma.

A side-effect is that in one existing test-case, the type is now parsed
as an ArrayOfBool instead of ArrayOfInt. The array is an array of
identifiers, so the parser does not identify the difference, and since
it attempts ArrayOfBool first, that is what it finds.